+++ iaso/tests/api/test_orgunits.py
@@ -387,131 +387,6 @@ class OrgUnitAPITestCase(APITestCase):
         self.assertEqual(1, m.BulkOperation.objects.count())
         self.assertEqual(3, am.Modification.objects.count())
 
-    @tag("iaso_only")
-    def test_org_unit_bulkupdate_task_select_all_but_some(self):
-        """POST /orgunits/bulkupdate/ happy path (select all except some)"""
-
-        from beanstalk_worker import task_service
-
-        task_service.clear()
-
-        self.client.force_authenticate(self.yoda)
-        response = self.client.post(
-            f"/api/orgunits/bulkupdate_task/",
-            data={
-                "select_all": True,
-                "validation_status": m.OrgUnit.VALIDATION_REJECTED,
-                "unselected_ids": [self.jedi_council_brussels.pk, self.jedi_council_endor.pk],
-                "groups_removed": [self.elite_group.pk],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-        data = response.json()
-        self.assertIn("task", data)
-        self.assertEqual(len(task_service.queue), 1)
-
-        task = Task.objects.get(pk=data["task"]["id"])
-        self.assertEqual(task.status, "QUEUED")
-        self.assertEqual(task.name, "org_unit_bulk_update")
-        self.assertEqual(task.launcher, self.yoda)
-
-        task_service.run_all()
-
-        self.jedi_council_corruscant.refresh_from_db()
-        self.assertEqual(self.jedi_council_corruscant.validation_status, m.OrgUnit.VALIDATION_REJECTED)
-        self.assertNotIn(self.elite_group, self.jedi_council_corruscant.groups.all())
-
-        for jedi_council in [self.jedi_council_endor, self.jedi_council_brussels]:
-            jedi_council.refresh_from_db()
-            self.assertTrue(jedi_council.validated)
-
-        self.assertEqual(1, m.BulkOperation.objects.count())
-        self.assertEqual(3, am.Modification.objects.count())
-
-        response = self.client.get("/api/tasks/%d/" % data["task"]["id"])
-
-        self.assertEqual(response.status_code, 200)
-        body = response.json()
-        self.assertEqual(body["status"], "SUCCESS")
-
-    @tag("iaso_only")
-    def test_org_unit_bulkupdate_task_select_all_with_multiple_searches(self):
-        """POST /orgunits/bulkupdate happy path (select all, but with multiple searches)"""
-        from beanstalk_worker import task_service
-
-        task_service.clear()
-        self.client.force_authenticate(self.yoda)
-        response = self.client.post(
-            f"/api/orgunits/bulkupdate_task/",
-            data={
-                "select_all": True,
-                "groups_added": [self.another_group.pk],
-                "searches": [{"validation_status": "all"}, {"validation_status": m.OrgUnit.VALIDATION_REJECTED}],
-            },
-            format="json",
-        )
-        self.assertJSONResponse(response, 200)
-        data = response.json()
-        self.assertIn("task", data)
-        self.assertEqual(len(task_service.queue), 1)
-
-        # check task
-        task = Task.objects.get(pk=data["task"]["id"])
-        self.assertEqual(task.status, "QUEUED")
-        self.assertEqual(task.name, "org_unit_bulk_update")
-        self.assertEqual(task.launcher, self.yoda)
-
-        task_service.run_all()
-        for jedi_council in [self.jedi_council_endor, self.jedi_council_brussels, self.jedi_council_corruscant]:
-            jedi_council.refresh_from_db()
-            self.assertIn(self.another_group, jedi_council.groups.all())
-
-        self.assertEqual(1, m.BulkOperation.objects.count())
-        self.assertEqual(5, am.Modification.objects.count())
-
-        response = self.client.get("/api/tasks/%d/" % data["task"]["id"])
-
-        self.assertEqual(response.status_code, 200)
-        body = response.json()
-        self.assertEqual(body["status"], "SUCCESS")
-
-    def test_task_kill(self):
-        """Launch the task and then kill it
-        Note this actually doesn't work if it's killwed while in the transaction part.
-        """
-
-        from beanstalk_worker import task_service
-
-        task_service.clear()
-        self.client.force_authenticate(self.yoda)
-        response = self.client.post(
-            f"/api/orgunits/bulkupdate_task/",
-            data={
-                "select_all": True,
-                "groups_added": [self.another_group.pk],
-                "searches": [{"validation_status": "all"}, {"validation_status": m.OrgUnit.VALIDATION_REJECTED}],
-            },
-            format="json",
-        )
-
-        self.assertJSONResponse(response, 200)
-        data = response.json()
-        self.assertIn("task", data)
-        self.assertEqual(len(task_service.queue), 1)
-        self.assertEqual(data["task"]["status"], "QUEUED")
-
-        task_model = Task.objects.get(id=data["task"]["id"])
-        task_model.should_be_killed = True
-        task_model.save()
-
-        task_service.run_all()
-
-        response = self.client.get("/api/tasks/%d/" % data["task"]["id"])
-
-        self.assertEqual(response.status_code, 200)
-        body = response.json()
-        self.assertEqual(body["status"], "KILLED")
 
     def test_org_unit_list_without_auth_or_app_id(self):
         """GET /api/orgunits/ with no auth or app id -> 200 with 0 org unit"""
