+++ iaso/urls.py
@@ -14,6 +14,7 @@ from .api.devices import DevicesViewSet
 from .api.devices_ownership import DevicesOwnershipViewSet
 from .api.devices_position import DevicesPositionViewSet
 from .api.data_sources import DataSourceViewSet
+from iaso.api.tasks.create.org_units_bulk_update import OrgUnitsBulkUpdate
 from .api.copy_version import CopyVersionViewSet
 from .api.dhis2_ou_importer import Dhis2OuImporterViewSet
 from .api.setup_account import SetupAccountViewSet
@@ -86,6 +87,7 @@ router.register(r"featureflags", FeatureFlagViewSet, basename="featureflags")
 router.register(r"copyversion", CopyVersionViewSet, basename="copyversion")
 router.register(r"dhis2ouimporter", Dhis2OuImporterViewSet, basename="dhis2ouimporter")
 router.register(r"setupaccount", SetupAccountViewSet, basename="setupaccount")
+router.register(r"tasks/create/orgunitsbulkupdate", OrgUnitsBulkUpdate, basename="orgunitsbulkupdate")
 router.register(r"tasks", TaskSourceViewSet, basename="tasks")
 
 router.registry.extend(plugins_router.registry)
